==> CrisisCorrelation/5.SPXAnalysis.qmd <==
---
title: "5.SPXAnalysis"
author: "Andrei Pazniak"
format: html
editor: visual
---

```{r init, echo=FALSE, include=FALSE}
library(data.table);library(ggplot2);library(plotly)
source("..\\common\\QuotesTransform.R")

data <- fread("cleanedSeries.csv")

targetName <- "^GSPC"

```

### TLT Price normal and adjusted

```{r}
p <- ggplot(data[series_id==targetName])  + geom_line(aes(x=date, y=adjusted_value_m2, color='Adjusted_M2 SP500')) +geom_line(aes(x=date, y=value, color="SP500")) + ggtitle("SP500 Price normal and adjusted") + ylab(targetName) + xlab("date")
ggplotly(p)
```

### Correlation Treasuries and FEDFUNDS

```{r}
data[series_id==targetName, y := growthIndex]
data[series_id=="FEDFUNDS", y := value/10]
d <- data[series_id==targetName | series_id == "FEDFUNDS", .(series_id, date, y)]

d2 <- d[series_id == "FEDFUNDS"]
d2[, series_id:="FEDFUNDS_MA200"]
d2[, y:= frollmean(y, 200, na.rm = T)]

d <- rbind(d, d2)


p <- ggplot(d, aes(x=date, y=y, color=series_id)) + geom_line() + ggtitle("SPX and FEDFUNDS") + ylab("growth Index/yield*0.1") + xlab("date")

ggplotly(p)

```

### FedFund trends

There are 3 states of FedFunds:

1.  Trend Up (If diffM200 for fedfunds is greater than threshold)

2.  Trend Down (If diffM200 for fedfunds is less than threshold)

3.  Sideway (If diffM200 for abs(fedfunds) is less than threshold)

To find threshold we make an assumption based on previous chart that 10% of time market in sideways and 90% in trend.

```{r}
fed_Sideways_Threshold <- quantile(data[series_id == "FEDFUNDS", abs(valueMVA-value)], probs=c(0.25), na.rm=TRUE)
print(paste("Threshold level for classification sideway is difference between moving average 200 and real value", fed_Sideways_Threshold[[1]]))

d_trend <- data[series_id == "FEDFUNDS", .(value, date, valueMVA)]
d_trend[value - valueMVA>=fed_Sideways_Threshold[[1]], trend:="up"]
d_trend[value - valueMVA<=-fed_Sideways_Threshold[[1]], trend:="down"]
d_trend[is.na(trend), trend:="sideway"]

d_sp <- data[series_id == targetName, .(value, date )]
d <- merge(d_sp, d_trend[, .(date, value_fed=value, trend)], by="date", all.x=T, all.y = F)


d_wide <- dcast(d, date  ~ trend, value.var = "value_fed")
d_long <- melt(d_wide, id.vars = c("date"), variable.name = "trend", value.name = "value")

p <- ggplot(d_long) + geom_line(aes(x=date, y=value, color=trend)) + geom_line(aes(x=date, y=value), alpha = 0.1) + ggtitle("Trends in FEDFUNDS") + ylab("growth Index/yield*0.1") + xlab("date")

ggplotly(p)
```

After trend classification we can see SPX \~ FED in convienient chart

```{r}
d_wide <- dcast(d, date  ~ trend)
d_long <- melt(d_wide, id.vars = c("date"), variable.name = "trend", value.name = "value")

p <- ggplot(d_long) + geom_line(aes(x=date, y=value, color=trend)) + geom_line(aes(x=date, y=value), alpha = 0.1) + ggtitle("SPX and FEDFUNDS") + ylab("price") + xlab("date")

ggplotly(p)
```

SP 500 growthIndex with colors

```{r}
d_sp <- data[series_id == targetName, .(growthIndex, date )]
d_growthIndex <- merge(d_sp, d_trend[, .(date, trend)], by="date", all.x=T, all.y = F)

d_wide <- dcast(d_growthIndex, date  ~ trend, value.var = "growthIndex")
d_long <- melt(d_wide, id.vars = c("date"), variable.name = "trend", value.name = "growthIndex")

p <- ggplot(d_long) + geom_line(aes(x=date, y=growthIndex, color=trend)) + geom_line(aes(x=date, y=growthIndex), alpha = 0.1) + ggtitle("SPX drawdown and FEDFUNDS") + ylab("drawdown") + xlab("date")

ggplotly(p)
```

todo refactor tp russel2000 \### TLT, Junk bond and their ratio

```{r bondyield}
bond <- data[date>'1998-01-01'][series_id=="BAMLH0A0HYM2EY" | series_id=="IRLTLT01USM156N"]
scaleFactor <- 10
d <- data.table(date = unique(bond$date), value = scaleFactor*bond[series_id=="IRLTLT01USM156N", value]/ bond[series_id=="BAMLH0A0HYM2EY", value], series_id = "tlt2junk", type="USA Economic/Percent")
d <- rbind(bond, d, fill=TRUE)
p <- ggplot(d, aes(x=date, y=value, color=series_id))  + geom_line() + geom_hline(yintercept = 2, color="blue") + geom_hline(yintercept = 5, color="blue") + ggtitle("TLT, Junk bond and their ratio") + ylab("yield/ratio") + xlab("date")
ggplotly(p)

```

Should we make 5 year investment to SPX when price is greater than MA200 or lower ? 
```{r scatterplot}
d<- data[series_id==targetName,]
#profit in 5 years
d[,profit_5y:= shift(value, 365*5, type="lead")/value]

dd <- 0.1

mean_random <- d[, mean(profit_5y, na.rm=TRUE)]
mean_belowMA200 <- d[diffM200<1-dd, mean(profit_5y, na.rm=TRUE)]
#AB testing to be sure that means are different
tt_result <- t.test(d[diffM200<1-dd, profit_5y], d[,profit_5y])
```

Analyse profit in next 5 years if buying when MA200 drawdown less then `dd*100`%. 
Confidence Interval is `tt_result$conf.int`

```{r ms200chart}
#just for research
#ggplot(d, aes(x=diffM200, y=profit_5y)) + geom_point() + geom_hline(yintercept=1, color="red") + #geom_vline(xintercept = 1, color="green") + ggtitle("Scatter plot 5Y profit ~ ma200 diff") 

#chart to understand how often situation with this condition
ggplot(d[diffM200<1-dd], aes(x=date, y=profit_5y)) + geom_col()
#comparison of buying instrument below MA200 on 10% or in any time 
ggplot() + 
  geom_density(data = d, aes(x = profit_5y, fill = "r"), alpha = 0.3) +
  geom_density(data = d[diffM200<1-dd], aes(x = profit_5y, fill = "b"), alpha = 0.3) +
  scale_colour_manual(name ="Comparison", values = c("r" = "red", "b" = "blue"), labels=c("b" = "Price < 0.x*MA200", "r" = "Random buying")) +
  scale_fill_manual(name ="Comparison", values = c("r" = "red", "b" = "blue"), labels=c("b" = "Price < 0.x*MA200", "r" = "Random buying")) +
  geom_vline(xintercept = 1, color="black") +
  geom_vline(xintercept = mean_random, color="red", ) +  
  geom_vline(xintercept = mean_belowMA200, color="blue") +
  annotate("text", x=mean_random, y=0.25, label="Mean Random Buying", size=3, color="red", angle = 90) + 
  annotate("text", x=mean_belowMA200, y=0.25, label="Mean BelowMA200 Buying", size=3, color="blue", angle = 90) +
  ylab("density")
```
Divide the GrowthChart into multiple waves, where each wave begins following 20% decline from an all-time high.

```{r improvedhist}
d <- data[series_id==targetName,.(date, series_id, value)]
d <- TranformToDropdawnData(d, columnName = "value")
alpha <- 0.2
TranformTo_LastMax(d, columnName = "value", threshold=1-alpha)
d[,`:=`(maxValue=NULL, dropdawn =NULL, dropdawnMVA =NULL, prevMax =NULL)]
wavesCuts <- as.POSIXct(c(d[1, date], d[growthIndex<=1-alpha & shift(growthIndex, 1)>1, date], d[.N, date]))
wavesData <- d[!is.na(value)][,waveId:=cut(as.POSIXct(date), breaks = wavesCuts, include.lowest = T)] 
wavesData[,index:=seq(.N), by=waveId]
wavesData[,dailyGrowth:=value/shift(value), by=waveId]
wavesData[!is.na(dailyGrowth),percentPrice:=cumprod(dailyGrowth), by=waveId]

ggplot(wavesData, aes(x=index, y=percentPrice, group=waveId, color=waveId)) + geom_line()
```
